requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
distlib==0.3.8
Django==3.2.25
django-cors-headers==4.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
filelock==3.12.2
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
importlib-metadata==6.7.0
oauthlib==3.2.2
platformdirs==4.0.0
proto-plus==1.24.0
protobuf==4.24.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.4
python-dotenv==0.21.1
pytz==2024.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.7
virtualenv==20.26.3
zipp==3.15.0
gunicorn==20.1.0

